LayoutTests/platform/wpe/imported/w3c/web-platform-tests/fetch/api/redirect/redirect-count-worker-expected.txt

CONSOLE MESSAGE: Load cannot follow more than 20 redirections
CONSOLE MESSAGE: Load cannot follow more than 20 redirections
CONSOLE MESSAGE: Load cannot follow more than 20 redirections
CONSOLE MESSAGE: Load cannot follow more than 20 redirections
CONSOLE MESSAGE: Load cannot follow more than 20 redirections

PASS Redirect 301 20 times
PASS Redirect 301 21 times
PASS Redirect 302 20 times
PASS Redirect 302 21 times
PASS Redirect 303 20 times
PASS Redirect 303 21 times
PASS Redirect 307 20 times
PASS Redirect 307 21 times
PASS Redirect 308 20 times
PASS Redirect 308 21 times
